feat: replaced individual menu  on CreateMenu-component
