+++ src/components/CreateMenu.js
@@ -4,9 +4,7 @@ import './CreateMenu.scss';
 
 import { Button } from 'react-bootstrap';
 
-import { Brekfast } from './creatingMenu/Brekfast';
-import { Lunch } from './creatingMenu/Lunch';
-import { Diner } from './creatingMenu/Diner';
+import { OneSet } from './creatingMenu/OneSet';
 
 function CreateMenu() {
     return (
@@ -28,10 +26,17 @@ function CreateMenu() {
                     </div>
                     {/* Menu selectors */}
                     <div className="CreateManu__MainBlock__Menu__MenuItem">
-                        <Brekfast />
-                        <Lunch />
-                        <Diner />
+                        <OneSet title="Сніданок" />
+                        <OneSet title="Обід" />
+                        <OneSet title="Вечеря" />
                     </div>
+                    {/* <div className="CreateManu__MainBlock__Menu__FullKcal">
+                        <Form>
+                            <Form.Label className="CreatingMenu__Menu__block__form__formItem">
+                                <Form.Control className="m-0" type="number" readOnly placeholder="калорійність" name="fullkll" value={formValues.fullKll} />
+                            </Form.Label>
+                        </Form>
+                    </div> */}
                     <Button className="CreateManu__MainBlock__Menu__SubButton" type="submit" variant="Success">
                         Зберегти
                     </Button>
